+++ saleor/plugins/vatlayer/tests/test_vatlayer.py
@@ -436,6 +436,23 @@ def test_apply_taxes_to_product(vatlayer, settings, variant, discount_info):
     assert price == TaxedMoney(net=Money("4.07", "USD"), gross=Money("5.00", "USD"))
 
 
+def test_apply_taxes_to_product_uses_taxes_from_product_type(
+    vatlayer, settings, variant, discount_info
+):
+    settings.PLUGINS = ["saleor.plugins.vatlayer.plugin.VatlayerPlugin"]
+    country = Country("PL")
+    manager = get_plugins_manager()
+    variant.product.metadata = {}
+    variant.product.product_type.metadata = {
+        "vatlayer.code": "standard",
+        "vatlayer.description": "standard",
+    }
+    price = manager.apply_taxes_to_product(
+        variant.product, variant.get_price([discount_info]), country
+    )
+    assert price == TaxedMoney(net=Money("4.07", "USD"), gross=Money("5.00", "USD"))
+
+
 def test_calculations_checkout_total_with_vatlayer(
     vatlayer, settings, checkout_with_item
 ):